Cast masked_image to VAE dtype before encoding in epicrealism inpainting
When the pipeline is cast to bfloat16, image_processor returns float32 tensors
which cause a dtype mismatch during VAE encoding.

+++ epicrealism_pureevolutionv5_inpainting/pytorch/src/model_utils.py
@@ -134,7 +134,7 @@ def inpainting_preprocessing(
 
     masked_image = image * (mask < 0.5)
     masked_image_latents = (
-        pipe.vae.encode(masked_image).latent_dist.mode()
+        pipe.vae.encode(masked_image.to(pipe.vae.dtype)).latent_dist.mode()
         * pipe.vae.config.scaling_factor
     )
